Version 12
This version is trained on a larger dataset of 35k (150x150) images. It took more than 2 hours to train on Nvidia GTX1650. UNet model (unet.py) does not have any skip connections or concatenations. Due to the lack of convolution layers after the transposed convolution, the output has checkerboard artifacts.